+++ sdk/storage/azure-storage-blob/src/samples/java/com/azure/storage/blob/BlobServiceClientJavaDocCodeSnippets.java
@@ -7,8 +7,10 @@ import com.azure.core.util.Context;
 import com.azure.storage.blob.models.BlobAnalyticsLogging;
 import com.azure.storage.blob.models.BlobContainerListDetails;
 import com.azure.storage.blob.models.BlobMetrics;
+import com.azure.storage.blob.models.BlobRequestConditions;
 import com.azure.storage.blob.models.BlobRetentionPolicy;
 import com.azure.storage.blob.models.BlobServiceProperties;
+import com.azure.storage.blob.options.ContainerRenameOptions;
 import com.azure.storage.blob.options.FindBlobsOptions;
 import com.azure.storage.blob.models.ListBlobContainersOptions;
 import com.azure.storage.blob.models.PublicAccessType;
@@ -318,5 +320,30 @@ public class BlobServiceClientJavaDocCodeSnippets {
             }
         );
         // END: com.azure.storage.blob.BlobServiceClient.undeleteBlobContainerWithResponse#UndeleteBlobContainerOptions-Duration-Context
+    }
+
+    /**
+     * Code snippet for {@link BlobServiceClient#renameBlobContainer(String, String)}
+     */
+    public void renameContainer() {
+        // BEGIN: com.azure.storage.blob.BlobServiceClient.createBlobContainer#String
+        BlobContainerClient blobContainerClient = client.renameBlobContainer("newContainerName", "oldContainerName");
+        // END: com.azure.storage.blob.BlobServiceClient.createBlobContainer#String
+    }
+
+    /**
+     * Code snippet for {@link BlobServiceClient#renameBlobContainerWithResponse(ContainerRenameOptions, Duration, Context)}
+     */
+    public void renameContainerWithResponse() {
+        // BEGIN: com.azure.storage.blob.BlobServiceClient.renameBlobContainerWithResponse#ContainerRenameOptions-Duration-Context
+        BlobRequestConditions requestConditions = new BlobRequestConditions().setLeaseId("lease-id");
+        Context context = new Context("Key", "Value");
+
+        BlobContainerClient blobContainerClient = client.renameBlobContainerWithResponse(
+            new ContainerRenameOptions("newContainerName", "oldContainerName")
+            .setRequestConditions(requestConditions),
+            Duration.ofSeconds(1),
+            context).getValue();
+        // END: com.azure.storage.blob.BlobServiceClient.renameBlobContainerWithResponse#ContainerRenameOptions-Duration-Context
     }
 }